タスクの管理は本人のみの実装

+++ src/services/task_service.ts
@@ -2,51 +2,51 @@ import { AppDataSource } from "../data-source";
 import Task from "../entity/Task";
 
 export const taskService = {
-  async findAll() {
+  async findAll(userId: number) {
     const repo = AppDataSource.getRepository(Task);
-    const allTasks = await repo.find();
+    const allTasks = await repo.find({ where: { userId } });
     return allTasks;
   },
 
-  async findOne(id: number) {
+  async findOne(id: number, userId: number) {
     const repo = AppDataSource.getRepository(Task);
-    const task = await repo.findOne({ where: { id } });
+    const task = await repo.findOne({ where: { id, userId } });
     return task;
   },
 
-  async create(content: string) {
+  async create(content: string, userId: number) {
     const repo = AppDataSource.getRepository(Task);
 
-    const task = repo.create({ content });
+    const task = repo.create({ content, userId });
     await repo.save(task);
 
     return task;
   },
 
-  async updateComplete(id: number, isCompleted: boolean) {
+  async updateComplete(id: number, isCompleted: boolean, userId: number) {
     const repo = AppDataSource.getRepository(Task);
 
     await repo.update(id, { isCompleted });
 
-    const task = await repo.findOne({ where: { id } });
+    const task = await repo.findOne({ where: { id, userId } });
 
     return task;
   },
 
-  async updateContent(id: number, content: string) {
+  async updateContent(id: number, content: string, userId: number) {
     const repo = AppDataSource.getRepository(Task);
 
     await repo.update(id, { content });
 
-    const task = await repo.findOne({ where: { id } });
+    const task = await repo.findOne({ where: { id, userId } });
 
     return task;
   },
 
-  async delete(id: number) {
+  async delete(id: number, userId: number) {
     const repo = AppDataSource.getRepository(Task);
 
-    const result = await repo.delete(id);
+    const result = await repo.delete({ id, userId });
 
     if (result.affected === 0) {
       return Error("Nothing task!");
